Move get_user_data to api.rs
